pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sales_guru"
version = "0.1.0"
description = "Sales Guru AI assistant"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "crewai",
    "crewai-tools",
    "jsonschema2md>=1.5.2",
    "litellm==1.55.0",
    "python-dotenv",
]

[project.scripts]
sales_guru = "sales_guru.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sales_guru"]

[tool.setuptools]
packages = ["sales_guru"]
package-dir = {"" = "src"}
